feat: SelectToppingSection ui 구현
